Rename file from data_types to types
