+++ src/parser/resp/mod.rs
@@ -4,7 +4,6 @@ pub(crate) mod big_number;
 pub(crate) mod boolean;
 pub(crate) mod bulk_error;
 pub(crate) mod bulk_string;
-pub(crate) mod data_types;
 pub(crate) mod double;
 pub(crate) mod integer;
 pub(crate) mod null;
@@ -12,4 +11,5 @@ pub(crate) mod simple_error;
 pub(crate) mod simple_string;
 
 // Exports
-pub use data_types::Type;
+pub(crate) mod types;
+pub use types::Type;
